extracted tests for Board, refactored count marks
