+++ src/board.py
@@ -2,11 +2,9 @@ class Board:
     def __init__(self):
         self.starter_board = ["1", "2", "3", "4", "5", "6", "7", "8", "9"]
 
-    def count_marks_in_board(self, player_one, player_two):
-        for mark in self.starter_board:
-            total_marks_on_board = self.starter_board.count(
-                player_one
-            ) + self.starter_board.count(player_two)
+    def count_marks_in_board(self, board, player_one, player_two):
+        for mark in board:
+            total_marks_on_board = board.count(player_one) + board.count(player_two)
         return total_marks_on_board
 
     def get_current_player(self, total_marks_on_board):
@@ -21,3 +19,9 @@ class Board:
         input_index = user_input - 1
         board[input_index] = self.get_current_player(total_marks_on_board)
         return board
+
+    def is_board_full(self, total_marks_on_board, board):
+        if total_marks_on_board == len(board):
+            return True
+        else:
+            return False
